(CLOUD-1640) Remove package resource for kubernetes-cni (#53)
Prior to this commit, kubernetes-cni was installed as a separate
package resource.  Regardless of the version specified, yum would
detect the kubelet prerequisite, which would automatically pull the
latest version.  Since we install kubelet just a few lines later,
and that installation will automatically pull the needed version
of kubernetes-cni, this package resource appears to not be needed,
and causes the incorrect version to get installed.

Removing it resolved the errors since the needed dependencies will be
pulled in when kubelet is installed.

would automatically pull in the most recent version