(CLOUD-1701) Move Exec resource defaults outside conditional (#57)
Prior to this commit, puppet agent runs on a controller would
fail because there is a relative path used in the assign master role
and taint master node execs.  These paths are present in the
resource default, but that is contained within the bootstrap-controller
conditional and thus not available for controllers.  This patch moves
the resource default above the conditional so it is present for both.