Allow mounting extra volumes to apiserver pod
In some environments, additional volumes are rqeuired for the apiserver
to function as expected. For example, on CentOS 7 controller nodes, the
host CA certificate bundles mounted at /etc/ssl/certs are broken since
this directory contains symlinks pointing to an unmounted path.

In other environments this may be used to provision a specific directory
which is referred to by other apiserver arugments, e.g.
"--oidc-ca-file".