Fix for failing idempotency on worker nodes
Fix for failing idempotency on worker nodes, these do not have the admin.conf.
Second run fails, because the exec resource has an unless which checks for nodes, which in turn fails because of the missing config file. This triggers an error, since the exec resource will attempt to join the node to the cluster for the second time.