Fix cni network provider (#227)
If a user deletes the flannel or the weave daemonset, a re-puppet should
add flannel back to the cluster. Because of the creates statement, one
would have to `rm -rf /etc/cni/net.d` and then re-puppet to get flannel
back. This is not intuitive behavor. This patch will re-apply flannel if
no flannel or weave daemonset exists.