Add default k8s conf directory creation for worker
Add resource to manage creation of the /etc/kubernetes default config
directory for worker nodes to ensure that similar to the controller
node it is always created before attempting to write any managed
configuration files. Test that the directory is a requirement of the
configuration file, as it should be added via autorequires thus
ensuring correct ordering.

First time manifest apply on a worker node will fail due to the config
file not being able to be written as the directory '/etc/kubernetes'
does not exist.

Subsequent runs do complete as the directory is created by a side
effect of another process being executed. However this prevents stand
up from pristine from succeeding.