Use correct repo location on Debian/Ubuntu
The kubernetes Debian repositories do not make use of codename-specific repositories. Instead a version for all supported releases is available in the "kubernetes-xenial" component. 

This change hardcodes this component name as the default accordingly. Of course this might brake in the future, but since the value can be overriden via a parameter and this change makes the default work on more distributions (e.g. Ubuntu Bionic) it seems to be a viable solution anyway.