(MAINT) Addressing wrong type for unless execs
Prior to this commit, as part of codebase hardening, unless commands
were wrapped in nested arrays so as to follow with proper convention.
However, a bug is found where unless commands containing a pipe | will
not work when broken down, nor when placed in a nested array.

This commit aims to revert some changes from the original codebase
hardening to address the failing conditionals.