(MAINT) Remove additional variables
This commit builds on the work that @deric submitted in a previous pr.

In a few other manifests there were variables set that added no value.
They have been reduced back to literal entries in the exec resources.

This commit applies those changes to this PR to keep things consistent.